+++ packages/fluentui/docs/src/prototypes/table/SelectableTable.tsx
@@ -10,23 +10,24 @@ import * as React from 'react';
 interface SelectableTableAction {
   type: 'TOGGLE_ITEM' | 'TOGGLE_ALL';
   checked: boolean;
-  itemIndex?: number;
+  itemKey?: string;
 }
 
 interface SelectableTableState {
   allRowsSelected: boolean;
-  selectedRows: boolean[];
+  selectedRows: {};
 }
 
 const selectableTableStateReducer: React.Reducer<SelectableTableState, SelectableTableAction> = (state, action) => {
   switch (action.type) {
     case 'TOGGLE_ITEM': {
-      const selectedRows = state.selectedRows.slice(0);
-      selectedRows[action.itemIndex] = action.checked;
-      return { allRowsSelected: !selectedRows.includes(false), selectedRows };
+      const selectedRows = { ...state.selectedRows };
+      selectedRows[action.itemKey] = action.checked;
+      return { allRowsSelected: !Object.values(selectedRows).includes(false), selectedRows };
     }
     case 'TOGGLE_ALL': {
-      const selectedRows = state.selectedRows.map(() => action.checked);
+      const selectedRows = { ...state.selectedRows };
+      Object.keys(selectedRows).forEach(key => (selectedRows[key] = action.checked));
       return { allRowsSelected: action.checked, selectedRows };
     }
     default:
@@ -36,9 +37,30 @@ const selectableTableStateReducer: React.Reducer<SelectableTableState, Selectabl
 
 const SelectableTable = () => {
   const rows = [
-    ['Test name 1', 'General', 'Office'],
-    ['Test name 2', 'Commander', 'Office'],
-    ['Test name 3', 'Other', 'Field'],
+    {
+      key: '1',
+      items: [
+        { key: '1-1', content: 'Test name 1' },
+        { key: '1-2', content: 'General' },
+        { key: '1-3', content: 'Office' },
+      ],
+    },
+    {
+      key: '2',
+      items: [
+        { key: '2-1', content: 'Test name 2' },
+        { key: '2-2', content: 'Commander' },
+        { key: '2-3', content: 'Office' },
+      ],
+    },
+    {
+      key: '3',
+      items: [
+        { key: '3-1', content: 'Test name 3' },
+        { key: '3-2', content: 'Surgeon' },
+        { key: '3-3', content: 'Field' },
+      ],
+    },
   ];
   const headers = [
     { title: 'Name', key: 'name', name: 'name' },
@@ -48,17 +70,19 @@ const SelectableTable = () => {
 
   const initialState: SelectableTableState = {
     allRowsSelected: false,
-    selectedRows: Array(rows.length).fill(false),
+    selectedRows: rows.reduce((selectedRows, row) => {
+      selectedRows[row.key] = false;
+      return selectedRows;
+    }, {}),
   };
   const [state, dispatch] = React.useReducer(selectableTableStateReducer, initialState);
 
   return (
     <>
       <Table aria-label="Selectable table" accessibility={gridNestedBehavior}>
-        <Table.Row header key="header" accessibility={gridRowBehavior}>
+        <Table.Row header accessibility={gridRowBehavior}>
           <Table.Cell
             accessibility={gridCellWithFocusableElementBehavior}
-            key="header-0"
             content={
               <Checkbox
                 title="Select all"
@@ -67,38 +91,35 @@ const SelectableTable = () => {
               ></Checkbox>
             }
           />
-          {headers.map((item, index) => (
-            <Table.Cell content={item.title} id={item.key} accessibility={gridCellBehavior} key={`header-${index}`} />
+          {headers.map(item => (
+            <Table.Cell content={item.title} accessibility={gridCellBehavior} key={item.key} />
           ))}
         </Table.Row>
-        {rows.map((row, index) => {
+        {rows.map(row => {
           return (
-            <Table.Row
-              key={`selectableTable${index}`}
-              accessibility={gridRowBehavior}
-              selected={state.selectedRows[index]}
-            >
+            <Table.Row accessibility={gridRowBehavior} selected={state.selectedRows[row.key]} key={row.key}>
               <Table.Cell
                 accessibility={gridCellWithFocusableElementBehavior}
-                key={`selectableTable${index}-0`}
                 content={
                   <Checkbox
                     title="Select me"
-                    checked={state.selectedRows[index]}
+                    checked={state.selectedRows[row.key]}
                     onClick={(event, props) =>
-                      dispatch({ type: 'TOGGLE_ITEM', checked: props.checked, itemIndex: index })
+                      dispatch({ type: 'TOGGLE_ITEM', checked: props.checked, itemKey: row.key })
                     }
                   ></Checkbox>
                 }
               />
-              {row.map((content, key) => (
-                <Table.Cell content={content} accessibility={gridCellBehavior} key={`selectableTable${index}-${key}`} />
+              {row.items.map(item => (
+                <Table.Cell content={item.content} accessibility={gridCellBehavior} key={item.key} />
               ))}
             </Table.Row>
           );
         })}
       </Table>
-      <Text content={`Selected rows: ${state.selectedRows.filter(Boolean).length}`} />
+      <Text
+        content={`Selected rows: ${Object.keys(state.selectedRows).filter(x => Boolean(state.selectedRows[x])).length}`}
+      />
     </>
   );
 };